# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Dive Hub API",
    "version": "0.3.0",
    "description": "Sign in with POST /api/auth/sign-in/email (Better Auth); the session cookie authenticates every other call."
  },
...
  "paths": {
...
    "/api/users/{id}": {
      "patch": {
        "summary": "Change a User's role; the last enabled admin stays admin (admins only)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "role"
                ],
                "properties": {
                  "role": {
                    "anyOf": [
                      {
                        "type": "string",
                        "enum": [
                          "user"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "admin"
                        ]
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "email",
                    "name",
                    "role",
                    "disabled",
                    "createdAt"
                  ],
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "role": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "user"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "admin"
                          ]
                        }
                      ]
                    },
                    "disabled": {
                      "type": "boolean",
                      "description": "Disabled Users can't sign in; their data stays"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "error"
                  ],
                  "properties": {
                    "code": {
                      "enum": [
                        "sign_in_required",
                        "admins_only",
                        "not_found",
                        "invalid_input",
                        "internal_error",
                        "upload_missing",
                        "upload_too_large",
                        "setup_done",
                        "setup_token_invalid",
                        "invitation_invalid",
                        "invitation_not_found",
                        "email_taken",
                        "reset_link_invalid",
                        "user_not_found",
                        "user_disabled",
                        "last_admin",
                        "not_yourself",
                        "confirmation_mismatch",
                        "session_not_found",
                        "import_not_found",
                        "dive_not_found",
                        "recording_not_found"
                      ],
                      "description": "Stable, machine-readable reason; clients translate it"
                    },
                    "error": {
                      "type": "string",
                      "description": "English description, for logs and scripts"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "error"
                  ],
                  "properties": {
                    "code": {
                      "enum": [
                        "sign_in_required",
                        "admins_only",
                        "not_found",
                        "invalid_input",
                        "internal_error",
                        "upload_missing",
                        "upload_too_large",
                        "setup_done",
                        "setup_token_invalid",
                        "invitation_invalid",
                        "invitation_not_found",
                        "email_taken",
                        "reset_link_invalid",
                        "user_not_found",
                        "user_disabled",
                        "last_admin",
                        "not_yourself",
                        "confirmation_mismatch",
                        "session_not_found",
                        "import_not_found",
                        "dive_not_found",
                        "recording_not_found"
                      ],
                      "description": "Stable, machine-readable reason; clients translate it"
                    },
                    "error": {
                      "type": "string",
                      "description": "English description, for logs and scripts"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "error"
                  ],
                  "properties": {
                    "code": {
                      "enum": [
                        "sign_in_required",
                        "admins_only",
                        "not_found",
                        "invalid_input",
                        "internal_error",
                        "upload_missing",
                        "upload_too_large",
                        "setup_done",
                        "setup_token_invalid",
                        "invitation_invalid",
                        "invitation_not_found",
                        "email_taken",
                        "reset_link_invalid",
                        "user_not_found",
                        "user_disabled",
                        "last_admin",
                        "not_yourself",
                        "confirmation_mismatch",
                        "session_not_found",
                        "import_not_found",
                        "dive_not_found",
                        "recording_not_found"
                      ],
                      "description": "Stable, machine-readable reason; clients translate it"
                    },
                    "error": {
                      "type": "string",
                      "description": "English description, for logs and scripts"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a User with everything only they own (admins only)",
        "description": "Removes their Originals and stored files, Imports, and the Divers no one else manages with all their Dives. Cannot be undone.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "confirmEmail"
                ],
                "properties": {
                  "confirmEmail": {
                    "type": "string",
                    "description": "The User's e-mail, typed by the admin to confirm"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "Default Response"
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "error"
                  ],
                  "properties": {
                    "code": {
                      "enum": [
                        "sign_in_required",
                        "admins_only",
                        "not_found",
                        "invalid_input",
                        "internal_error",
                        "upload_missing",
                        "upload_too_large",
                        "setup_done",
                        "setup_token_invalid",
                        "invitation_invalid",
                        "invitation_not_found",
                        "email_taken",
                        "reset_link_invalid",
                        "user_not_found",
                        "user_disabled",
                        "last_admin",
                        "not_yourself",
                        "confirmation_mismatch",
                        "session_not_found",
                        "import_not_found",
                        "dive_not_found",
                        "recording_not_found"
                      ],
                      "description": "Stable, machine-readable reason; clients translate it"
                    },
                    "error": {
                      "type": "string",
                      "description": "English description, for logs and scripts"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "error"
                  ],
                  "properties": {
                    "code": {
                      "enum": [
                        "sign_in_required",
                        "admins_only",
                        "not_found",
                        "invalid_input",
                        "internal_error",
                        "upload_missing",
                        "upload_too_large",
                        "setup_done",
                        "setup_token_invalid",
                        "invitation_invalid",
                        "invitation_not_found",
                        "email_taken",
                        "reset_link_invalid",
                        "user_not_found",
                        "user_disabled",
                        "last_admin",
                        "not_yourself",
                        "confirmation_mismatch",
                        "session_not_found",
                        "import_not_found",
                        "dive_not_found",
                        "recording_not_found"
                      ],
                      "description": "Stable, machine-readable reason; clients translate it"
                    },
                    "error": {
                      "type": "string",
                      "description": "English description, for logs and scripts"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "error"
                  ],
                  "properties": {
                    "code": {
                      "enum": [
                        "sign_in_required",
                        "admins_only",
                        "not_found",
                        "invalid_input",
                        "internal_error",
                        "upload_missing",
                        "upload_too_large",
                        "setup_done",
                        "setup_token_invalid",
                        "invitation_invalid",
                        "invitation_not_found",
                        "email_taken",
                        "reset_link_invalid",
                        "user_not_found",
                        "user_disabled",
                        "last_admin",
                        "not_yourself",
                        "confirmation_mismatch",
                        "session_not_found",
                        "import_not_found",
                        "dive_not_found",
                        "recording_not_found"
                      ],
                      "description": "Stable, machine-readable reason; clients translate it"
                    },
                    "error": {
                      "type": "string",
                      "description": "English description, for logs and scripts"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/users/{id}/disable": {
      "post": {
        "summary": "Block sign-in and end all sessions; data stays (admins only)",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "email",
                    "name",
                    "role",
                    "disabled",
                    "createdAt"
                  ],
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "role": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "user"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "admin"
                          ]
                        }
                      ]
                    },
                    "disabled": {
                      "type": "boolean",
                      "description": "Disabled Users can't sign in; their data stays"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "error"
                  ],
                  "properties": {
                    "code": {
                      "enum": [
                        "sign_in_required",
                        "admins_only",
                        "not_found",
                        "invalid_input",
                        "internal_error",
                        "upload_missing",
                        "upload_too_large",
                        "setup_done",
                        "setup_token_invalid",
                        "invitation_invalid",
                        "invitation_not_found",
                        "email_taken",
                        "reset_link_invalid",
                        "user_not_found",
                        "user_disabled",
                        "last_admin",
                        "not_yourself",
                        "confirmation_mismatch",
                        "session_not_found",
                        "import_not_found",
                        "dive_not_found",
                        "recording_not_found"
                      ],
                      "description": "Stable, machine-readable reason; clients translate it"
                    },
                    "error": {
                      "type": "string",
                      "description": "English description, for logs and scripts"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "error"
                  ],
                  "properties": {
                    "code": {
                      "enum": [
                        "sign_in_required",
                        "admins_only",
                        "not_found",
                        "invalid_input",
                        "internal_error",
                        "upload_missing",
                        "upload_too_large",
                        "setup_done",
                        "setup_token_invalid",
                        "invitation_invalid",
                        "invitation_not_found",
                        "email_taken",
                        "reset_link_invalid",
                        "user_not_found",
                        "user_disabled",
                        "last_admin",
                        "not_yourself",
                        "confirmation_mismatch",
                        "session_not_found",
                        "import_not_found",
                        "dive_not_found",
                        "recording_not_found"
                      ],
                      "description": "Stable, machine-readable reason; clients translate it"
                    },
                    "error": {
                      "type": "string",
                      "description": "English description, for logs and scripts"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "error"
                  ],
                  "properties": {
                    "code": {
                      "enum": [
                        "sign_in_required",
                        "admins_only",
                        "not_found",
                        "invalid_input",
                        "internal_error",
                        "upload_missing",
                        "upload_too_large",
                        "setup_done",
                        "setup_token_invalid",
                        "invitation_invalid",
                        "invitation_not_found",
                        "email_taken",
                        "reset_link_invalid",
                        "user_not_found",
                        "user_disabled",
                        "last_admin",
                        "not_yourself",
                        "confirmation_mismatch",
                        "session_not_found",
                        "import_not_found",
                        "dive_not_found",
                        "recording_not_found"
                      ],
                      "description": "Stable, machine-readable reason; clients translate it"
                    },
                    "error": {
                      "type": "string",
                      "description": "English description, for logs and scripts"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/users/{id}/sessions": {
      "delete": {
        "summary": "Sign a User out everywhere (admins only)",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "Default Response"
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "error"
                  ],
                  "properties": {
                    "code": {
                      "enum": [
                        "sign_in_required",
                        "admins_only",
                        "not_found",
                        "invalid_input",
                        "internal_error",
                        "upload_missing",
                        "upload_too_large",
                        "setup_done",
                        "setup_token_invalid",
                        "invitation_invalid",
                        "invitation_not_found",
                        "email_taken",
                        "reset_link_invalid",
                        "user_not_found",
                        "user_disabled",
                        "last_admin",
                        "not_yourself",
                        "confirmation_mismatch",
                        "session_not_found",
                        "import_not_found",
                        "dive_not_found",
                        "recording_not_found"
                      ],
                      "description": "Stable, machine-readable reason; clients translate it"
                    },
                    "error": {
                      "type": "string",
                      "description": "English description, for logs and scripts"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "error"
                  ],
                  "properties": {
                    "code": {
                      "enum": [
                        "sign_in_required",
                        "admins_only",
                        "not_found",
                        "invalid_input",
                        "internal_error",
                        "upload_missing",
                        "upload_too_large",
                        "setup_done",
                        "setup_token_invalid",
                        "invitation_invalid",
                        "invitation_not_found",
                        "email_taken",
                        "reset_link_invalid",
                        "user_not_found",
                        "user_disabled",
                        "last_admin",
                        "not_yourself",
                        "confirmation_mismatch",
                        "session_not_found",
                        "import_not_found",
                        "dive_not_found",
                        "recording_not_found"
                      ],
                      "description": "Stable, machine-readable reason; clients translate it"
                    },
                    "error": {
                      "type": "string",
                      "description": "English description, for logs and scripts"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "error"
                  ],
                  "properties": {
                    "code": {
                      "enum": [
                        "sign_in_required",
                        "admins_only",
                        "not_found",
                        "invalid_input",
                        "internal_error",
                        "upload_missing",
                        "upload_too_large",
                        "setup_done",
                        "setup_token_invalid",
                        "invitation_invalid",
                        "invitation_not_found",
                        "email_taken",
                        "reset_link_invalid",
                        "user_not_found",
                        "user_disabled",
                        "last_admin",
                        "not_yourself",
                        "confirmation_mismatch",
                        "session_not_found",
                        "import_not_found",
                        "dive_not_found",
                        "recording_not_found"
                      ],
                      "description": "Stable, machine-readable reason; clients translate it"
                    },
                    "error": {
                      "type": "string",
                      "description": "English description, for logs and scripts"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}